# lib/api-zod/src/generated/types/competitor.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sieve marketing intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface Competitor {
  id: number;
  brandId: number;
  name: string;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
}