# lib/api-zod/src/generated/types/memoResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sieve marketing intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { MemoResponseRunType } from "./memoResponseRunType";
import type { MemoSections } from "./memoSections";
import type { SourceRef } from "./sourceRef";

export interface MemoResponse {
  id: number;
  runType: MemoResponseRunType;
  /** @nullable */
  query?: string | null;
  /** @nullable */
  rationale_summary?: string | null;
  /** @nullable */
  confidence?: number | null;
  sections: MemoSections;
  sourceRefs: SourceRef[];
  status: string;
  createdAt: Date;
}