# lib/api-zod/src/generated/types/playbookStep.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sieve marketing intelligence API
 * OpenAPI spec version: 0.1.0
 */

export interface PlaybookStep {
  id: number;
  playbookId: number;
  stepOrder: number;
  stepTitle: string;
  /** @nullable */
  stepDescription?: string | null;
}