# lib/api-zod/src/generated/types/principle.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Sieve marketing intelligence API
 * OpenAPI spec version: 0.1.0
 */
import type { PrincipleStatus } from "./principleStatus";

export interface Principle {
  id: number;
  title: string;
  statement: string;
  /** @nullable */
  explanation?: string | null;
  domainTag: string;
  /** @nullable */
  confidenceScore?: string | null;
  sourceCount: number;
  sourceRefsJson: string;
  status: PrincipleStatus;
  createdAt: Date;
  updatedAt: Date;
}